+++ benches/bench_compare_la16.rs
@@ -1,7 +1,6 @@
 use std::num::NonZeroU32;
 
 use glassbench::*;
-use image::imageops;
 
 use fast_image_resize::pixels::U16x2;
 use fast_image_resize::{CpuExtensions, FilterType, Image, MulDiv, PixelType, ResizeAlg, Resizer};
@@ -10,28 +9,10 @@ use testing::PixelExt;
 mod utils;
 
 pub fn bench_downscale_la16(bench: &mut Bench) {
-    let src_image = U16x2::load_big_image().to_luma_alpha16();
     let new_width = NonZeroU32::new(852).unwrap();
     let new_height = NonZeroU32::new(567).unwrap();
 
     let alg_names = ["Nearest", "Bilinear", "CatmullRom", "Lanczos3"];
-
-    // image crate
-    // https://crates.io/crates/image
-    for alg_name in alg_names {
-        let filter = match alg_name {
-            "Nearest" => imageops::Nearest,
-            "Bilinear" => imageops::Triangle,
-            "CatmullRom" => imageops::CatmullRom,
-            "Lanczos3" => imageops::Lanczos3,
-            _ => continue,
-        };
-        bench.task(format!("image - {}", alg_name), |task| {
-            task.iter(|| {
-                imageops::resize(&src_image, new_width.get(), new_height.get(), filter);
-            })
-        });
-    }
 
     // fast_image_resize crate;
     let mut cpu_ext_and_name = vec![(CpuExtensions::None, "rust")];
@@ -49,12 +30,12 @@ pub fn bench_downscale_la16(bench: &mut Bench) {
                 "Lanczos3" => ResizeAlg::Convolution(FilterType::Lanczos3),
                 _ => return,
             };
-            let src_image_data = U16x2::load_big_src_image();
-            let src_view = src_image_data.view();
+            let src_image = U16x2::load_big_src_image();
+            let src_view = src_image.view();
             let mut premultiplied_src_image = Image::new(
-                NonZeroU32::new(src_image.width()).unwrap(),
-                NonZeroU32::new(src_image.height()).unwrap(),
-                src_view.pixel_type(),
+                src_image.width(),
+                src_image.height(),
+                src_image.pixel_type(),
             );
             let mut dst_image = Image::new(new_width, new_height, PixelType::U16x2);
             let mut dst_view = dst_image.view_mut();
@@ -71,9 +52,7 @@ pub fn bench_downscale_la16(bench: &mut Bench) {
             bench.task(format!("fir {} - {}", ext_name, alg_name), |task| {
                 task.iter(|| match resize_alg {
                     ResizeAlg::Nearest => {
-                        fast_resizer
-                            .resize(&premultiplied_src_image.view(), &mut dst_view)
-                            .unwrap();
+                        fast_resizer.resize(&src_view, &mut dst_view).unwrap();
                     }
                     _ => {
                         mul_div